pyproject.toml
[tool.poetry]
name = "netrics-measurements"
version = "0.0.1"
description = "The extensible network measurements framework"
license = "MIT"
authors = [
  "Jesse London <jesselondon@gmail.com>",
  "Kyle MacMillan <macmillan@uchicago.edu>",
]
packages = [{include = "netrics", from = "src"}]

[tool.poetry.dependencies]
python = "^3.8"
#
# FIXME: See dssg/argcmdr#27
#
fate-scheduler = {url = "https://github.com/chicago-cdac/fate/archive/0.0.2.post6.zip"}

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
netrics = "netrics:main"

# built-in measurement modules
netrics-dev = "netrics.measurement.dev:main"
netrics-dns-latency = "netrics.measurement.dns_latency:main"
netrics-ndt7 = "netrics.measurement.ndt7:main"
netrics-ookla = "netrics.measurement.ookla:main"
netrics-ping = "netrics.measurement.ping:main"
netrics-traceroute = "netrics.measurement.traceroute:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
